protocol/resp: flush once per command
This commit overrides `Async::IO::Protocol::Line#write_line`.

The original method performs a flush. This one does not and moves the
responsibility of flushing to the caller of the method.

In the case of Redis, we do not want to perform a flush in every line,
because each Redis command contains several lines. Flushing once per
command is more efficient because it avoids unnecessary writes to the
socket.